# src/main/java/com/tenniscourts/guests/GuestDTO.java
package com.tenniscourts.guests;

import javax.validation.constraints.NotNull;
import lombok.AllArgsConstructor;
import lombok.Builder;
import lombok.Getter;
import lombok.NoArgsConstructor;
import lombok.Setter;

@AllArgsConstructor
@NoArgsConstructor
@Builder
@Getter
@Setter
public class GuestDTO {

    private Long id;

    @NotNull
    private String name;
}